Increase email send cronjob timeout to 1h
It seems to take approximately half an hour in practice to send
notifications to all children belonging to one project (year). It
should not be fatal if one cronjob timeouts (because the next one
should continue where the previous left), but it is of course better if
there are no timeouts, thus this change.