Fix free spot subscription handling in event group
Enrolling to an occurrence should delete all free spot notification
subscriptions from the child which are for the same event, or for an
event that is in the same event group as the enrolled occurrence.
Before this change, subscriptions were not deleted from the other
events of the event group.