Add initial report API
Added a separate REST API at URL `/reports/` which provides anonymized
data to be used for reporting purposes. There is currently one endpoint,
`/children/`, which returns all of the children with the exception of
"orphan" ones, as we don't consider those valid cases ATM.

To access the API, one must use HTTP basic auth (or be logged in when
`DEBUG` is `True`), and the user must also have Django permission
`reports.access_report_api`.