Add feedback notification
Added a new automatic notification which is meant to be sent after an
occurrence has ended to ask feedback from the participants. The
notifications are sent by a management command which needs to be run
periodically in staging and production. There is also a configurable
additional delay with a default value of 15min.