Implement child.activeInternalAndTicketSystemEnrolments GraphQL query
Implemented query for fetching upcoming and ongoing (with
leeway) internal ticket system enrolments (Enrolment model instances)
and external ticket system "enrolments" (assigned TicketSystemPassword
instances). The enrolments are sorted from earliest to latest where time
is either the related occurrence's time or the related event's latest
occurrence's time depending on which kind of enrolment it is.

The current purpose of this query is to provide a way for the UI to
fetch a child's all upcoming enrolments to "Tulevat tapahtumasi" (your
upcoming events) section in child profile view.