Add ticket system URL to EventNode.ticketSystem in API
Added Event field `ticket_system_url` to EventNode.ticketSystem object
in the GraphQL API. It can be queried and mutated.

Also removed the ability to change ticket system type on event update,
because that ability wasn't needed or used, and it would made the event
updation logic related to the new field more complex for no reason.