Add ticket system end time field to Event model
Added field `ticket_system_end_time` field to Event model. The purpose
for the field is to tell when a Ticketmaster event ends so that it can
be returned in child's past events. The field is needed because
Ticketmaster events don't have occurrences which are used to determine
end times for internal ticket system events.