Process notifications sending outside the atomic transaction of saving of events
KK-984.
The object instance save-process should be handled in a transaction,
so if any errors occurs, the whole process is rolled back.
The sending of notifications (emails and sms) cannot be in that same transaction,
because a sent message cannot be rolled back --
the receiver has already received the message.

The notifications context should be processed as much as possible
before the actual sending, so that any issues should have been already raised,
before any of the messages has been delivered yet.