Add Lippupiste as an external ticket system
KK-987.

The integration for the external ticket systems,
the Ticketmaster and the Lippupiste, should be similar.
Only thing that varies at this point , is the ticket system URL.

The old Ticketmaster model and schema classes are now renamed
and converted to some abstract classes. The old Ticketmaster class names
are now given to the new classes that inherits the abstracts.
New Lippupiste classes are copied from the new Ticketmaster ones.

The old implementation for Ticketmaster should be unchanged,
so all the existing unit tests should pass without any changes.
There are now some changes added there only to randomly test that also
the Lippupiste events are calculated in same manners.