test: pagination and offset of the Children query
KK-1049.

Add a test that tests the pagination of the Children-query with a larger
amount of data compared to what has earlier been tested with.

Add a test that checks that the offset and after parameters usage responds
with the same results.

Add test for pagination cursor generator.

There has been issues handling the pages after 2nd page,
when there has been data for around 70 children
and the page size has been set to something that results in more than 2 pages.