feat!: rename the birthdate field of the child model to be birthyear
KK-1024.

The date of the birth is not needed, but the year is.

The plan:
The birthdate field is first renamed to be birthyear.
After field name change, the field type is converted from
a date to an integer. Data migration is needed to be ran first,
so that we don't lose the birth year that is already filled
within the date.