feat: add query and mutation for guardian marketing subscriptions
KK-1072.

A new query, MyMarketingSubscriptions, added along side with
a new mutation, UpdateMyMarketingSubscriptions.
They both can be used with a auth verification token,
so they do not mandatorily require an authentication through a login form,
but an auth key can be used as an argument instead.

A new fields to the GraphQL endpoint were introduced,
because the subscriptions handling needs to be available
also when the user cannot login, so that they can remove
their email subscription when they cannot authenticate anymore.
This is a compromise in security, so instead allowing my profile updates with
a token, a new query and mutation was added with less information available.