feat: user can unsubscribe from all the notifications at once
KK-1072.

There are some new functions added to User model that can be used to
make the unsubscribing to all the notification subscriptions.

Added a method to `FreeSpotNotificationSubscriptionQueryset` to easily
filter the user's free sport notification subscriptions. A child instance can be used
as an argument to filter the result set even further.