fix: upgrade the django-helusers to fix issues with GDPR API auth
KK-1097 KK-1117.

When the GDPR API requests data from the Kukkuu API,
it must use an authorization token (which in this point is issued by Tunnistamo).
The issued JWT auth token has the amr-field in a string format, 
when it should be a list. Upgrading the helusers-package fixes the auth issues.