feat: send event and event group publish notification in a thread
KK-1035.

Added a new common decorator that can be used to execute tasks in background
in a new thread.

If there are thousands of email recipients, the process of creating the mails
will take tens of seconds. While the process is handled in the same
request handler as the event state change, the response takes long to be sent,
which leads to a connection timeout, that will write an error message
to the admin UI. To prevent that, the state change could be done as an atomic
operation as it has always been done, but the emails could be sent in an another
thread. The response is then sent when the atomic process finishes and the email
sending is not awaited.

More about the threading for email sending:
- https://ozcanyarimdunya.github.io/dj_threading/
- https://medium.com/@akshatgadodia/simplifying-email-sending-in-django-a-step-by-step-guide-7dc90079738b
fix: db connection was not working properly in pytests

Some of the pytests needed some threading handling: the database connection was not shared
with the newly spawned threads, so e.g. the sent mails were not shown in the mail box because
the querysets were empty. Using `@pytest.mark.django_db(transaction=True)` helped in that.