feat: event report API with event group and venue view sets as helpers
KK-1026.

Added 3 new API view sets:
1. The most important one: the Events report API. An optimised list view that is mostly used by PowerBI reports.
2. Event Group list and detail view, to offer more details for events report API.
3. Venue list and detail view, to offer more details for events report API.

Multiple new serializers were needed. Serializers for
1. Event
2. Occurrence
3. EventGroup
4. Project
5. Venue

Optimize the serializers so that they don't fetch trivial data for Power BI reports.
The list view does not need all the translations or calculations, since they can be offered
from the details/retrieve view. For example, it does not make sense for PowerBI to read the name
multiple times from every event group that is linked to event. So the name of the event group can be
excluded from the events list view and be rendered only in the event group details/retrieve view.