fix: get_translations_dict utility's database usage
KK-1026.

The `get_translations_dict` is meant to be used to fetch all the translations for 1 field at once and return them in a dictionary.
When investigated, it was clear that it didn't respect the `fetch_related('translations')` call from the manager,
but did always make a new query for each and every field that was translated for every item in a result list.

With the added fix, it now respects the prefetch_related-command and the queries are much better optimized.