feat(auth): authentication with graphene_jwt instead of custom implementation
KK-1224.

Install `graphene_jwt` and use it's Graphene middleware instead of our
own custom one. Also removed the custom middleware for auth and error
handling.

Because of that and since the own custo middleware was removed, the
kukkuu.oidc `authenticate` needed changes: instead of returning an
UserAuthorization, the actual User instances was needed to be returned.

NOTE: The graphene_jwt's middleware checks that the request must have an
authorization header with a bearer token.