feat: project message permissions for sending to all
KK-1039.

Add a project permission that defines whether a user can send a message
to all members in a project. The new project permission will be
available to be selected in project object permissions, and in user and
group admin forms.

Also refactored the other existing project permissions so that they are
all listed in a new enum. The new enum helps the future refactoring of
the code, because it is now a referable constant.