fix: remove ChildNode deprecation warnings and document fields' behavior
Remove the deprecation warnings from:
 - `ChildNode.available_events`
 - `ChildNode.available_events_and_event_groups`
 - `EventQuerySet.available`

Add note to description of ChildNode fields that they don't take yearly
enrolment limits into account:
 - `available_events`
 - `available_events_and_event_groups`
 - `upcoming_events_and_event_groups`

NOTE:
 - The added descriptions of ChildNode fields are shown in the UIs'
   generated types as the fields' comments, the generated types just
   need to be regenerated to reflect this

So basically this change is about not deprecating the fields' current
behavior but documenting it. As the fields in question are in use and
have been in use for some time it means their functionality has been
apparently "good enough". If their functionality needs to be changed,
then that could be ticketed, or a TODO/FIXME added.

refs KK-1357