refactor(auditlog-extra): override get_objects in admin view mixin
KK-1113.

Instead of overriding all the admin views separately, override just the
get_object function, because it is called to fetch a single object (when
ever admin site reads details of only 1 object) and it is also used by
3rd party plugins. Overriding the get_object is much more effective that
overriding just couple of views offered by Django. Also, there is a risk
that the views might change on new releases, when the get_object is much
more resilient.

A use of contextvar is needed to prevent the accessed signal to be
called multiple times.