fix(csp): circular import in settings
KK-1360.

When the CSP class is added to the constants and is used in
`settings.py`, the `DEFAULT_LANGUAGE` in `consts.py` made a circular
import. It's better if the constants does not have any imports from
anywhere and since the DEFAULT_LANGUAGE is the same as
`settings.LANGUAGE_CODE`, it is easy to remove the DEFAULT_LANGUAGE.