chore(fixture): provide mandatory projects with fixture
KK-1394.

Remove the existing data migration that added a Project instance for the
year 2020.

Include all the projects from year 2020 to 2024 in a new projects
fixture JSON file.

The data is dumped with `python manage.py dumpdata --natural-primary`
from the test environment's service pod. Then the data keys are managed
so that the primary keys makes more sense. The data should be imported
with Django's `python manage.py loaddata projects`.