fix: update projects to db using migrations and management commands
KK-1394.

Add a new data migration file that adds the missing mandatory projects,
1 per year after year 2020. When children are added to the system, they
always needs to be linked to a year project of their birth year.

Update the tests, because the new data migration will add more projects
to the database and some of the tests were trusting in 1 specific state.

Fix the Project model `__str__` with a safe translation getter, since
the management command has not set any language by default.