fix: updating requirements by moving non-PyPI packages to own .txt file
`pip-compile requirements.in` didn't work, it failed with RemoteNotFound
but that was fixed by adding the /opt/app-root/src/django-auditlog-extra
directory to git's safe directories in Dockerfile's development stage:
```
RUN git config --system --add safe.directory \
    /opt/app-root/src/django-auditlog-extra
```

After that `pip-compile requirements.in` worked but then
`pip-compile requirements-dev.in` failed with:
```
PipDeprecationWarning: DEPRECATION: Constraints are only allowed to take
the form of a package name and a version specifier. Other forms were
originally permitted as an accident of the implementation, but were
undocumented. The new implementation of the resolver no longer supports
these forms. A possible replacement is replacing the constraint with a
requirement. Discussion can be found at
https://github.com/pypa/pip/issues/8210
```

So basically using -e / editable installs i.e. packages outside of PyPI
can not be used as constraints currently and they need to be handled in
some other way.

This commit handles the only editable package install currently used
i.e. django-auditlog-extra package by installing it using
`requirements-not-from-pypi.txt` without using it as a constraint.

refs KK-1409