feat: remove is_obsolete field from user model
KK-1251.

The `is_obsolete` field in User model was only used during the migration
process when authentication service was changed from Tunnistamo to a
Keycloak service of Helsinki Profile. It was deprecated after all the
users were (data) migrated so that all the obsoleted users were also
deactivated. There is no need for user.is_obsolete anymore, so it could
be deleted.